refactor: remove navigationExtensions
This commit removes navigation extensions.It was removed because, the
extension function defined in it allowed to only call navigate and did
not have overloads that were present in the default navigation lib.
This caused the source code to use this extension function in some areas
and to use the standard navigate() function in other areas.To prevent
in-consistencies, this extension, along with the file in which it was defined in, was removed.