fix(PocketRepository): unmark previously deleted item if saveUrlForUser is called with the same URL
Previously, a check if an item already exists in the database was made
before the item was saved. If it did exist PocketRepository#saveUrlForUser
would just return without performing any operation. This caused an issue
because it prevented the user from saving the same item again. This commit
adds a change that checks whether the url already exists. If it does, it
then checks if it is marked as deleted. If it is marked as deleted the
item will be unmarked. Else, the method will just return.